Refactor out postPersistRequest to utils
